Updated weather-mv BQ Metadata.
Rows now contain a record of the path to the input dataset from their original ingestion. They also each contain the first time step, which (by convention) is the same as the time when a forecast was made.

Both of these new fields can be used to differentiate duplicate forecasts from different time steps, in order to select the most accurate prediction.