Increasing workers per key for MARS to documented limit. (#19)
* Increasing workers per key for MARS to documented limit.

* Updated to best estimate of current limits for MARS and CDS.

Pipeline honors client limits; makes no other changes.

* Fixed lint issue: extra space.

* Added pipeline option to specify number of threads per worker.

* Updated constants in tests.

* Updated docstrings with a helpful description.
Includes steps to cancel open requests.

* Ran e2e; fixed type cast error.

* Improved algorithm for configuring concurrent workers.

Added ability for user to specify number of requests per license. Then, we compute the number of workers to allocate, depending on the number of threads and user configuration. If the user overrides the max number of threads and will likely hit an error, we provide a warning.

* Implementing feedback: better copy.

Co-authored-by: Alex Merose <alxrsngrtn@google.com>