Updating config assembly to evenly distribute licenses / keys.
Here, we make partition assembly with subsections a stateful process. This helps us keep track of recently used licenses and ensures that keys are evenly distributed – despite randomly dispatched workers.

In addition to this main contribution, this PR cleans up the overall flow of the download pipeline to be simpler and more testable.