New data-oriented task distribution strategy. (#116)
* New Data-Oriented Task distribution strategy.

In this change, I divide up the download requets to workers in a data-driven way. Namely, after fanning out the partitions, I group by both the license and number of requests. Then, I've restructured the fetch step to execute each group of configs in order.

This way, we can guarantee that all licenses are being utilized to their max capacity and within rate limits. This is a much better fix for #98, as it also obviates the needs to limit the max number of workers (this is actually autoscale friendly!)

Thanks to zyda@google.com for the idea and an example for how to implement this.

* Removed configure workers method.

Now that we have a data-oriented strategy for evenly distributing tasks, we don't need to fight autoscaling anymore. Thus, we can get rid of the code that sets the max number of workers.

* Incrementing weather-dl version.

* Better logging in the fetch step.

* Nit: Better counter key.

* Fix: assert + indentation error.

* Improved readability of `parse_target_name`.

* Parser tests are parameterized.

* A simpler, more parallel version of the pipeline.