Downloader refactor: extracted out partitioning; tested pipeline args.
This change improves the modularity and testability of the downloader pipeline. Here, we extract out the partition preparation steps into its own composite PTransform. This provides a better location for the partitioning logic, and makes testing how we partition much better.

Next, we've decoupled running the pipeline from processing user arguments, and added unit tests to the user processing part. This ensures we can exercise logic about how to configure the pipeline without actually running it.