Robust download from clients to VMs. (#143)
This fixes #140 with a few tactics. First, I've added a check to skip a partition during the fetch step. Thus, if this step fails and restarts, it won't re-download from the beginning anymore. This functions as a fallback mechanism. Second, we add retry logic that will catch a common source of error: socket timeouts during download from remote servers. This should prevent the ParDo step from failing in the first place.

Thanks to @uhager for identifying this bug.