Unscheduling developer meetings.
The resourcing for our project has changed; we have more internal contributors at this time than external. As a result, we plan on coordinating asynchronously with external contributors. In our experience over the past month, we found that meetings over video-chat were not that fruitful.

We can revisit reinstating these open meetings at a future time.