Fixed potential memory leak in grib splitting.
When we open a grib locally, we currently do not use a context manager to open it. This means that `close()` is never called on the open grib data, which would thus leak memory.

In addition to this fix, I've refactored a bit of pipeline code to use less variables.