Parallelizing fan-out step during `PreparePartitions` in `weather-dl`.  (#236)
In this PR, we change the prepare partitions steps to a) use indexes instead of generating most of the configs and b) to split partitions (i.e. cartesian cross-products) into lists of lists. We aim to divide all partitions into a 50 lists (when the number is big), and offer users a -p, --partition-chunks argument to let users influence the number of groups.