Multi-config support in `weather-dl`, with in-order and 'fair' scheduling. (#270)
Fixes #9. In addition, this CL gives the user the ability to set how downloads are scheduled. Specifically, the user can choose if partitions will be ordered in the same order of input configs, or if they should be evenly distributed from each config in a "fair" way.

* Naive implementation of multiple-config support.

* fix typos in configs.

* Implemented fair scheduling.

* Failing fewer tests...

* Not sure how to get the last test to pass.

* Fixed last failing test.

* Documented the default behavior.

* Better partition defaults for fair scheduling.

* Updated scheduling to better support multiple licenses per config.

I got my math wrong and didn't account for configs with multiple licenses. This commit hopes to change that. Thanks @mahrsee1997 for pointing this out!

* Updated license to correct year.

* Removed accidental comment.

* Updated scheduling math to account for number of requests per key.

* Updated tests.

* Feedback from @mahrsee1997.