Log warning instead at failure to parse projection info with rasterio. (#348)
We use rasterio on open datasets to gather projection information and the datatype of data opened with `open_dataset()`. In cases where rasterio (or GDAL) cannot parse the URI, let's log a warning instead of crashing the pipeline.

When the CRS info is needed (in the `weather-mv ee` pipeline), users can pass this in via a custom COG conversion transform.